more e2e restructuring

+++ lambda/downloader/service.py
@@ -1,8 +1,14 @@
 import os
 import boto3
+import logging
+from lib import memory
 from lib.util import parse_sns_event
 from lib.models import Image, ImageUsage, Session
 from lib.image import image_hash, download_image, downscale_image
+from requests.exceptions import HTTPError
+
+logger = logging.getLogger()
+logger.setLevel(logging.INFO)
 
 DOWN_SIZE = (800, 800)
 HASH_SIZE = 16 # must be a power of 2. 16 for more accuracy, but slower
@@ -27,31 +33,47 @@ def handler(event, context):
     # and compute the image hash from the original.
     # we don't save the original because it will be too expensive.
     url = event.get('url')
-    img = download_image(url)
-    hash = str(image_hash(img, HASH_SIZE))
-    key = hash
+    if not memory.already_downloaded(url):
+        try:
+            img = download_image(url)
+        except Exception as e:
+            logger.error('Exception {} : {}'.format(e, url))
+            if isinstance(e, HTTPError) and e.response.status_code == 404:
+                session.close()
+                return
+            else:
+                raise
+        hash = str(image_hash(img, HASH_SIZE))
+        key = hash
 
-    meta = {
-        'height': img.height,
-        'width': img.width,
-        'format': img.format,
-        'mode': img.mode
-    }
+        meta = {
+            'height': img.height,
+            'width': img.width,
+            'format': img.format,
+            'mode': img.mode
+        }
 
-    # check if image already is saved
-    image = session.query(Image).filter(
-        Image.hash==hash,
-        Image.url==url
-    ).first()
-    if image is None:
-        # save the downscaled image
-        out = downscale_image(img, DOWN_SIZE, QUALITY)
-        bucket.put_object(Key=key, Body=out, ACL='public-read')
+        # check if image already is saved
+        image = session.query(Image).filter(
+            # assuming that images are unique to urls,
+            # for now
+            # Image.hash==hash,
+            Image.url==url
+        ).first()
+        if image is None:
+            # save the downscaled image
+            out = downscale_image(img, DOWN_SIZE, QUALITY)
+            bucket.put_object(Key=key, Body=out, ACL='public-read')
 
-        # create new image
-        image = Image(url=url, key=key, hash=hash, **meta)
-        session.add(image)
-        session.commit()
+            # create new image
+            image = Image(url=url, key=key, hash=hash, **meta)
+            session.add(image)
+            session.commit()
+        memory.record_downloaded(url)
+    else:
+        image = session.query(Image).filter(
+            Image.url==url
+        ).first()
 
     # associate the image with the context
     context_id = event.get('context_id')
